+++ x/incentive/keeper/keeper_shares.go
@@ -1,0 +1,47 @@
+package keeper
+
+import (
+	"math"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// Calculate total share of staking
+func (k Keeper) CalculateTotalShareOfStaking(amount sdk.Int) sdk.Dec {
+	// Total statked = Elys staked + Eden Committed + Eden boost Committed
+	totalStaked := k.tci.TotalElysBonded.Add(k.tci.TotalEdenEdenBoostCommitted)
+	if totalStaked.LTE(sdk.ZeroInt()) {
+		return sdk.ZeroDec()
+	}
+
+	// Share = Amount / Total Staked
+	return sdk.NewDecFromInt(amount).QuoInt(totalStaked)
+}
+
+// Calculate the delegated amount
+func (k Keeper) CalculateDelegatedAmount(ctx sdk.Context, delegator string) sdk.Int {
+	// Derivate bech32 based delegator address
+	delAdr, err := sdk.AccAddressFromBech32(delegator)
+	if err != nil {
+		// This could be validator address
+		return sdk.ZeroInt()
+	}
+
+	// Get elys delegation for creator address
+	delegatedAmt := sdk.ZeroDec()
+
+	// Get all delegations
+	delegations := k.stk.GetDelegatorDelegations(ctx, delAdr, math.MaxUint16)
+	for _, del := range delegations {
+		// Get validator address
+		valAddr := del.GetValidatorAddr()
+		// Get validator
+		val := k.stk.Validator(ctx, valAddr)
+
+		shares := del.GetShares()
+		tokens := val.TokensFromSharesTruncated(shares)
+		delegatedAmt = delegatedAmt.Add(tokens)
+	}
+
+	return delegatedAmt.TruncateInt()
+}
